% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_phenotype_function.R
\name{plot.pRRopheticCv}
\alias{plot.pRRopheticCv}
\title{Plot "pRRopheticCv" object.}
\usage{
\method{plot}{pRRopheticCv}(cvOutData)
}
\arguments{
\item{cvOutData}{an object of class "pRRopheticCv", i.e. the outpout of the "predictionAccuracyByCv()" function}
}
\description{
Given an object of class "pRRopheticCv", plot the cross validation
predicted values against the measured values. Also plots a regression
line.
}
\keyword{plot}

