\name{summary.pRRopheticCv}
\alias{summary.pRRopheticCv}
\title{Summary of "pRRopheticCv" object.}
\usage{
\method{summary}{pRRopheticCv}(cvOutData)
}
\arguments{
  \item{cvOutData}{an object of class "pRRopheticCv", i.e.
  the outpout of the "predictionAccuracyByCv()" function}
}
\description{
Given an object of class "pRRopheticCv", print various
metrics that summarize the performance of the cross
validataion analysis
}
\keyword{summary}

